[chore][refactor] Aplicando clean arch
