+++ src/main/java/br/com/vrsoftware/adapters/auth/IAuthenticationStrategy.java
@@ -1,0 +1,7 @@
+package br.com.vrsoftware.adapters.auth;
+
+import java.net.http.HttpRequest;
+
+public interface IAuthenticationStrategy {
+    void applyAuthentication(HttpRequest.Builder requestBuilder);
+}
