--- src/main/java/br/com/vrsoftware/application/usecases/interfaces/IJiraIssueRepository.java
@@ -1,8 +1,0 @@
-package application.usecases.interfaces;
-
-import domain.entities.JiraIssue;
-import java.util.List;
-
-public interface IJiraIssueRepository {
-    JiraIssue getById(String id);
-}
