--- src/main/java/br/com/vrsoftware/domain/entities/jira/ActorDTO.java
@@ -1,39 +1,0 @@
-package br.com.vrsoftware.dto.jira;
-
-import br.com.vrsoftware.exceptions.SerializationException;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class ActorDTO {
-
-    String displayName;
-    String emailAddress;
-
-    public ActorDTO() {
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public void setDisplayName(String displayName) {
-        this.displayName = displayName;
-    }
-
-    public String getEmailAddress() {
-        return emailAddress;
-    }
-
-    public void setEmailAddress(String emailAddress) {
-        this.emailAddress = emailAddress;
-    }
-
-    @Override
-    public String toString() {
-        try {
-            return new ObjectMapper().writeValueAsString(this);
-        } catch (JsonProcessingException e) {
-            throw new SerializationException("Falha ao serializar Actor.", e);
-        }
-    }
-}
