+++ src/main/java/br/com/vrsoftware/exceptions/AuthenticationException.java
