+++ src/main/java/br/com/vrsoftware/exceptions/EncryptionException.java
