+++ src/main/java/br/com/vrsoftware/exceptions/RequestException.java
