+++ src/main/java/br/com/vrsoftware/exceptions/SerializationException.java
