--- src/main/java/br/com/vrsoftware/service/auth/IAuthenticationStrategy.java
@@ -1,7 +1,0 @@
-package br.com.vrsoftware.service.auth;
-
-import java.net.http.HttpRequest;
-
-public interface IAuthenticationStrategy {
-    void applyAuthentication(HttpRequest.Builder requestBuilder);
-}
