+++ src/main/java/br/com/vrsoftware/usecases/ObjectMapperConfig.java
@@ -1,0 +1,30 @@
+package br.com.vrsoftware.usecases;
+
+import br.com.vrsoftware.exceptions.SerializationException;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class ObjectMapperConfig {
+
+    private static final ObjectMapper MAPPER = new ObjectMapper()
+            .setSerializationInclusion(JsonInclude.Include.NON_NULL)
+            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+
+    public static String toJson(Object obj) {
+        try {
+            return MAPPER.writeValueAsString(obj);
+        } catch (JsonProcessingException e) {
+            throw new SerializationException("Falha ao serializar objeto.", e);
+        }
+    }
+
+    public static <T> T fromJson(String content, Class<T> valueType) {
+        try {
+            return MAPPER.readValue(content, valueType);
+        } catch (JsonProcessingException e) {
+            throw new SerializationException("Falha ao deserializar objeto.", e);
+        }
+    }
+}
